test: split the parallel-round suite out of api-handlers
api-handlers.test.ts crossed the 500-line cap; the `add --count` block was
already a self-contained suite (the old fan-out one), so it moves to
api-add-parallel.test.ts with the parallel flag guards that belong beside
it. Same 33 tests.
